infrastructure: modify prompt
When you have a million terminals open it can be nice easily know which
of the terminals is in the Zephyr environment. This this patch adds a
few lines in the zephyr-env.sh and zephyr-env.cmd files to modify the
prompt when running source zephyr-env.sh or calls the zephyr-env.cmd.
I really don't know how write scripts for the command line in Windows,
but I think the code in zephyr-env.cmd should be correct.

Signed-off-by: Even Falch-Larsen <even.falch-larsen@nordicsemi.no>