[nrf noup] drivers: serial: uarte: fix UARTE driver
When the TX interrupt was disabled, the UARTE driver was stopping
the transmission abruptly, thus garbling the output; wait for the
last transmission to complete after disabling the TX irq instead.

The NRF_UARTE_EVENT_ENDTX was not being cleared when the application
expected it to be cleared. When checking for uart_irq_tx_ready()
also check that TX interrupts are enabled.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>