[nrf noup] cmake: multi image build support with IMAGE prefix
Note: This was originally submitted upstream in #13672 and was
      rejected. NCS is carrying it as a noup for now. Another,
      more upstream-friendly approach will be needed.

We are seeing the need to boot Zephyr applications in multiple
stages, a real-world use-case of even a 3-stage bootloader has
been identified and tested.

Each bootloader needs to be individually upgrade-able and have
it's own configurations of Zephyr libraries. To achieve this each
bootloader is organized as it's own executable.

The problem is that the build system can only build one
executable. This creates a usability issue as the user must invoke a
large set of arcane commands to build, sign, and flash each
bootloader.

To resolve this we re-organize the build system such that it can build
multiple executables. To work within CMake restrictions that logical
target names must be globally unique (among other naming
restrictions), we add an IMAGE variable which is used to name the
current Zephyr application being built.

Signed-off-by: Sebastian Bøe <sebastian.boe@nordicsemi.no>
Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>
Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>
Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>
Signed-off-by: Ruth Fuchss <ruth.fuchss@nordicsemi.no>

(cherry picked from commit 3db22acba35aa62abb18f4823102401f8b211a57)
(Conflicts were resolved in lib/posix/CMakeLists.txt. Changes to
 ia32.cmake were dropped entirely since Intel has rejected this
 approach.

 Finally, these follow-up fixes and dependent patches were squashed
 in:

  02eeeb779665f126f59c185430d854bc7392f18e
  c6c3fed492e0f49679f5c6230dfc955c72f4ffaf
  56f6a33a71c46ec5b2e1beb6076edf9521014ad3
  1dc0ca5ccead37a53f600c65a941938f4138cf38
  6e348b8e609109981cbc140f06c324d938b52cd4)
Signed-off-by: Marti Bolivar <marti.bolivar@nordicsemi.no>