[nrf fromlist] cmake: zephyr_modules: Add west executable argument
PR [#15666](https://github.com/zephyrproject-rtos/zephyr/pull/15666)

On some systems where you don't have access to `PATH` and you can't
set the `ENV{PATH}` variable. You need to be able to pass the path
to the west executable down to the python script so it is better
for it to be set explicitly than assuming that it exsists as a
part of the PATH/executables in the shell being called.

Signed-off-by: Sigvart M. Hovland <sigvart.hovland@nordicsemi.no>