cmake: depend on file AND wrapper target for merged_to_sign
In some corner cases, it is required to depend on both
the wrapper target for a file generated through add_custom_command
AND the file being generated.

Prior to this fix, the dependency from 'merged_to_sign.hex'
to 'signed.hex' was not being enforced. This resulted in
outdated 'merged.hex' files when re-building.

This commit fixes this dependency, making a re-build
correctly re-generate the 'signed.hex' and hence 'merged.hex'.

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>