[nrf noup] scripts: partition_manager: Split into two files
One file for the actual algorithm. The file takes a number of yaml files
and outputs a processed yaml file.

The new file contains all the other file generation (header and kconfig)
It takes the output yaml as input.

Define and use a new GLOBAL PROPERTY PM_IMAGES which contains only the
image names, instead of the '|'-separated struct. The rest of the info
that was previously in the struct is now inferred from the image name
and the new ${IMAGE}PROJECT_BINARY_DIR property.

Create only one instance of files that are identical for all images:
 - partitions.yml
 - pm.config
Place these in the root folder.

Change the input parameters to the script to better fit with the
changes to the build flow.

Signed-off-by: Øyvind Rønningstad <oyvind.ronningstad@nordicsemi.no>