[nrf toup] cmake: imageify OVERLAY_CONFIG
Currently what is provided in -DOVERLAY_CONFIG will be applied
to all images, this is incorrect. This patch ensures that the
overlay config is only applied to a single image.

Note that it is also possible to specify which image it should
apply to by providing the image name as a prefix in the command.

E.g. '-Dspm_OVERLAY_CONFIG'

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>