cmake: fix load offset of mcuboot built for s1
The 'PM_S0_SIZE' property used is incorrect, as this includes padding
which is before the 's0_image' partition, which mcuboot is originally
linked against.

Fix this by using 'PM_S0_IMAGE_SIZE' property instead, which does not
contain the size of the pad. See bootloader/pm.yml for details.

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>