[nrf noup] cmake: adjust offset after introducing alignment
Calculate the offset needed for S1-linked mcuboot based on address
of S1 and S0. This change is needed since empty partitions
can be injected between any partition when alignment is used.
If an empty partition was inserted between S0 and S1 the old way
of calculating the offset would be incorrect.

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>