[nrf fromlist] Bluetooth: host: Fix out of range parameters [...]
provided to the controller

Fix LE Create Connection command giving out of range parameters to the
controller, this came back as 0x30 (Out of Range Parameters) status code
on the command from the controller.
This appears to be the min and max CE parameters in the command.
Revert back memset from 137f70406411

Submitted upstream as:
https://github.com/zephyrproject-rtos/zephyr/pull/22374

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>