[nrf noup] cmake: change how path to subsys pm.yml is set
The relative path from the build dir to the binary dir of a subsys
folder is not static. In a zephyr subsys it is

build/zephyr/subsys/path

whereas in a nrf subsys it is

build/modules/nrf/subsys/path

The previous way of setting the path to preprocessed pm.yml files
for subsystems only worked with the first of these.

This commit introduces the use of a cmake variable for the binary dir,
and a mechanism for setting the build dir path for each pm.yml
explicitly.

Signed-off-by: Håkon Øye Amundsen <haakon.amundsen@nordicsemi.no>
(cherry picked from commit 79f1bb595339a6ba0f281c4beb3fec79da817f0d)
Signed-off-by: Andrzej Puzdrowski <andrzej.puzdrowski@nordicsemi.no>