[nrf noup] subsys/fs: move FS partition to the end
In case of the DFU capability enabled:
Up to NCS v1.1.0 any storage partition lands at the end of
the flash memory.
Since v1.2.0 any storage partition lands right after
the executable application.
So it is impossible to build proper update application out-of-the box.

This patch moves storage partitions at the flash end back.
User needs only to correct the partition size (via Kconfig) to
the same as in previous release for ensure proper upgrade build.

Signed-off-by: Andrzej Puzdrowski <andrzej.puzdrowski@nordicsemi.no>