[nrf noup]: using ZEPHYR_NRF_MODULE_DIR for partition manager inclusion
Now using ZEPHYR_NRF_MODULE_DIR when including partition manager cmake
file.
The previous OPTIONAL loading of partition manager would include the
file if the nrf folder was present regardless of whether nrf is part
of the project or not.

This means a stray nrf folder causes unintended side effects when using
sdk-zephyr as manifest repository.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>