[nrf temphack] drivers: flash: lock around read/write pair
A write of sub-word data requires reading the content currently at the
target offset so that a word-sized write doesn't change it.
Correctness requires no other thread be given an opportunity to change
that data before the subsequent write is performed.

Extend the wait API to allow a lock to be retained after a failed
operation.  Rework the write control flow to allow success in the read
to continue to the write without unlocking, while failure in the read
bypasses the write.

Signed-off-by: Peter A. Bigot <pab@pabigot.com>