[nrf fromtree] Bluetooth: L2CAP: Elevate security level when connect ..
request denied

Elevate the security level of the connection when the l2cap channel
connect request is denied with encryption or authentication error
code.
The l2cap channel required_sec_level was used, but this is not
correct as this is set by the user of the API. This resulted in
the call to bt_conn_set_security seeing that the current security
level was already set, decide there is nothing to be done and return
success.

Fixes: #27220

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
(cherry picked from commit 9c3d0064712247475d21245d64c2357c84428678)
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>