[nrf fromtree] kernel/timeout: revert to signed dticks for timeout storage
The fix to handle long duration timeouts in 150e18de6e1806942 also
changed the duration value from signed to unsigned.  This can cause
delays in processing alarms when timer handlers run longer than a tick
(either due to delays or if there are many of them).  Revert to a
signed representation.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>