[nrf temphack] drivers: ieee802154: nrf5: Remove tx_started callback
This is a fixup to commit:
851eac4ae2d [nrf fromlist]
drivers: ieee802154: nrf5: Add support for tx in the future

and should be reverted when
https://github.com/zephyrproject-rtos/zephyr/pull/29123
is upmerged.

Remove the implementation of nrf_802154_tx_started
radio driver callback. Remove the semaphore dedicated
to this callback.

Signed-off-by: Maciej Fabia <maciej.fabia@nordicsemi.no>
(cherry picked from commit 1d4bbe5b7695f003b823c4d307706efdd4a44415)