[nrf fromtree] twisterlib: checking for classname and testname
Currently duplicated testcases are only determined by the classname.
This results in testcases not beeing added to the twister.xml if the
testcases are from the same class (e.g. sample.yaml), but with different
names.

The changed check if the combination of classname and testname
already exists.

Upstream commit: 148dbdc9fefc53a58c1fe990ebfc9390cb4a09f6

Signed-off-by: Sebastian Wezel <sebastian.wezel@nordicsemi.no>