[nrf fromtree] Bluetooth: host: Refactor update_keys_check to operate ..
on keys as input

Refactor update_keys_check helper function to operate on input keys
input. This allows the function to be re-used on a keys structure that
is not the current connection keys.

This also avoids the helper function changing the connection state.
The conn->le.keys pointer should at this point always have been
assigned, as central when sending the pairing request, and as peripheral
when receiving the pairing request at the very latest.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
(cherry picked from commit 7fee6b8ae835426a6322168582f4ad9e53162b87)
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>