[nrf fromtree] Bluetooth: host: Overwrite existing bond when IRK has ...
been updated

Overwrite the existing bond when the IRK of the existing bond could not
resolve the RPA of the peer. This would happen if the peer has deleted
the bond and replaced the IRK that was used.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
(cherry picked from commit e06bac0e4f2537ec1547c1c2cca80d5f25fdf0a3)
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>