[nrf fromlist] cmake: zephyr_file() now accepts BOARD and BOARD_REVISION
... as argument

With the introduction of #32556 changing the BOARD now prints a warning.

Zephyr provides `zephyr_file()` to look up overlays and Kconfig
fragments for a specified board, and it used to be possible and safe to
do:
```
function(my_func)
  set(BOARD <local_scope_board>)
  zephyr_file(CONF_FILES ...)
endfunction(my_func)
```

As the BOARD inside `my_func` is locally scoped, this is safe to do.
But with introduction of #32556 a warning is now printed when running
CMake.

Therefore `zephyr_file` has been extended to allow for optional BOARD
and BOARD_REVISION arguments.
If BOARD is not given as argument the current BOARD in the build system
will be used.

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/33373

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>