[nrf fromlist] tests: bluetooth: tester: Remove assignment to ...
... CONFIG_LOG_MINIMAL

The LOG_MINIMAL option is no longer user-configurable, as now it has
no prompt. LOG_MODE_MINIMAL should be modified instead.
But in this case, the assignment is redundant, as this option is not
enabled by default. Remove it then, to align the configuration file
for nrf5340dk_nrf5340_cpuapp with the one for nrf52840dk_nrf52840.

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/33460

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>