[nrf fromtree] kernel: fix error in synchronous work cancellation
...return value

The return value is documented to be true if the work was pending, but
the implementation returned true only if the work was actually running
(i.e. the caller had to wait).  It should also return true if
scheduled or submitted work was cancelled.

Note that this means the return value cannot be used to determine
whether the call slept.

Signed-off-by: Peter Bigot <peter.bigot@nordicsemi.no>
(cherry picked from commit 707dc22fb06a60a69329c04a299a30d36fb160eb)
Signed-off-by: Trond Einar Snekvik <Trond.Einar.Snekvik@nordicsemi.no>