[nrf fromtree] boards: mps2_an521: remove dependencies for
...BUILD_WITH_TFM

We do not need to have dependencies any more on !OPENAMP
for BUILD_WITH_TFM in the mps2 an521 target, because we
now have different targets for the non-secure version of
the board and the remote core.

[nrf note]: This is already cherry-picked once in #511, but due to
reversal of a temphack, it was only partially picked, and the changes in
this patch were missing.

Signed-off-by: Ioannis Glaropoulos <Ioannis.Glaropoulos@nordicsemi.no>
(cherry picked from commit e04a57de046e0e1fba25d3ecc11bd6d7611e3664)
Signed-off-by: Trond Einar Snekvik <Trond.Einar.Snekvik@nordicsemi.no>