[nrf fromtree] drivers: serial: nrf_uart: Disable uart when fifo_fill not used
Driver was failing when interrupt mode was enabled for given
instance but interrupt driven TX part was not used. In that
case uart was not disabled after sending a byte which resulted
in continuous interrupt triggering. Added check for
fifo_fill_lock which is set when uart_fifo_fill is used.

Signed-off-by: Krzysztof Chruscinski <krzysztof.chruscinski@nordicsemi.no>
(cherry picked from commit 485f593bb75b8228cb2ef5bc1debdceaddfb5c32)
Signed-off-by: Andrzej Puzdrowski <andrzej.puzdrowski@nordicsemi.no>