[nrf fromtree] Bluetooth: Mesh: Don't reset PB ADV reliable timer on
...retransmit

The send_reliable function was reused in multiple places as part of the
k_delayed_work changes for Bluetooth Mesh in #33782. This function
contains a line that resets the start timer, causing prov_retransmit to
continously move the goal post for when to give up sending.

Extract this line out of the send_reliable function, and put it along
with the other link.tx initialization in bearer_ctl_send and
prov_send_adv.

(cherry picked from commit 85ad497c39ecf6d12bea31241d774cf300d620c2)
Signed-off-by: Trond Einar Snekvik <Trond.Einar.Snekvik@nordicsemi.no>