[nrf fromtree] samples: drivers: watchdog: Pause during debugging
Enable the option to pause the watchdog if the MCU is halted by a
debugger.

This fixes an issue with some Nordic MCUs (see #33509) where the board
could not be flashed anymore if a short watchdog timeout (<100 ms) was
used.

Signed-off-by: Martin Jäger <martin@libre.solar>

cherry-picked from 784e6847df2c81e203b03dfbf2bbc0456439cdf9

Signed-off-by: Maciej Perkowski Maciej.Perkowski@nordicsemi.no