[nrf fromtree] Bluetooth: host: Fix extended advertising feature bit ...
check

Fix CONFIG_BT_EXT_ADV_LEGACY_SUPPORT option which optimizes the host
to skip checking the feature bit of the controller to check for
extended advertising commands.
This was broken because of how this was implented using an undef of the
feature bit, which was not replicated in scan.c, adv.c and id.c once
this was split out from hci_core.

Instead of doing this wierd way of redefining the feature check macro
do it in a proper way by defining a new macro.

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>
(cherry picked from commit b6c95edc639e48fadf8be4bfd18c95839baf2ab1)
Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>