[nrf fromlist] soc: nrf53: Fix configuration of HFXO capacitance
This is a follow-up to commit 60d9988401e59093edb0cd638b1b737288477cb0.

Add a third option for the HFXO capacitance that keeps the default
value of the XOSC32MCAPS register untouched. The message in the above
commit incorrectly claimed that external load capacitors for HFXO
(the reset value of the XOSC32MCAPS register) was the configuration
in effect before. In fact, the register value was modified during
the device trimming in the SystemInit() function to use the internal
capacitors, and that is the configuration required for proper RADIO
operation on nRF5340 DK, for instance.

Upstream PR: https://github.com/zephyrproject-rtos/zephyr/pull/36320

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>