[nrf noup] REVERTME: ci: Bump to use CMake 3.20.5
NOTE TO MAINTAINERS:

   This should be reverted before the next upmerge.

   It was originally cherry picked from commit
   e78cb747bd850eacd8008155937ddf9426102984 ("ci: Bump to use CMake
   3.20.5") upstream, but manual conflict resolution was required.
   Rather than bring in other additional patches, we're just
   replicating its effect here in NCS for CI purposes. We'll get it
   back from upstream after reverting when we next synchronzie.

With CMake minimum required as 3.20.0 we update CI to use docker image
v0.18.2, which contains CMake 3.20.5.

For doc builds we fetch the same CMake v3.20.5 but using pip as the doc
build doesn't use the docker image.

The main reason for increasing CMake version is better toolchain
support.
The decision to bump the CMake version was taken by the Toolchain WG.

Better toolchain support is added in the following CMake versions:
- armclang, CMake 3.15
- Intel oneAPI, CMake 3.20
- IAR, CMake 3.15 and 3.20

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>
Signed-off-by: Martí Bolívar <marti.bolivar@nordicsemi.no>