[nrf fromlist] drivers: flash: Make `mock_flash` public
For getting the address of the RAM region for nRF53 DFU we need the
`mock_flash` to be public so that we can import it via `extern`. This is
because the PCD interface of the nRF53 Network Core DFU procedure
requires a size and start address. Where the start address when
emulating the primary slot will be in RAM we need some way of extracting
this information.

Upstream PR: zephyrproject-rtos/zephyr#37783

Signed-off-by: Sigvart Hovland <sigvart.m@gmail.com>